:recycle: refactor(api): move `query-params-utils.js` to `src`
