+++ api/lib/application/target-profiles/target-profile-controller.js
@@ -3,7 +3,7 @@ import { usecases as devcompUsecases } from '../../../src/devcomp/domain/usecase
 import * as targetProfileSerializer from '../../infrastructure/serializers/jsonapi/target-profile-serializer.js';
 import * as targetProfileSummaryForAdminSerializer from '../../infrastructure/serializers/jsonapi/target-profile-summary-for-admin-serializer.js';
 import * as targetProfileForAdminSerializer from '../../infrastructure/serializers/jsonapi/target-profile-for-admin-serializer.js';
-import * as queryParamsUtils from '../../infrastructure/utils/query-params-utils.js';
+import * as queryParamsUtils from '../../../src/shared/infrastructure/utils/query-params-utils.js';
 import * as organizationSerializer from '../../infrastructure/serializers/jsonapi/organization-serializer.js';
 import * as badgeSerializer from '../../../src/shared/infrastructure/serializers/jsonapi/badge-serializer.js';
 import { deserializer as badgeCreationDeserializer } from '../../infrastructure/serializers/jsonapi/badge-creation-serializer.js';
