+++ api/src/prescription/organization-learner/application/learner-list-controller.js
@@ -1,7 +1,7 @@
 import { usecases } from '../domain/usecases/index.js';
 import * as organizationParticipantsSerializer from '../infrastructure/serializers/jsonapi/organization-participants-serializer.js';
 import { mapCertificabilityByLabel } from './../../shared/application/helpers.js';
-import * as queryParamsUtils from '../../../../lib/infrastructure/utils/query-params-utils.js';
+import * as queryParamsUtils from '../../../shared/infrastructure/utils/query-params-utils.js';
 
 const findPaginatedFilteredParticipants = async function (
   request,
