+++ api/tests/prescription/organization-learner/unit/application/learner-list-controller_test.js
@@ -3,7 +3,7 @@ import { expect, hFake, sinon, domainBuilder } from '../../../../test-helper.js'
 import { usecases } from '../../../../../src/prescription/organization-learner/domain/usecases/index.js';
 import { learnerListController } from '../../../../../src/prescription/organization-learner/application/learner-list-controller.js';
 
-import * as queryParamsUtils from '../../../../../lib/infrastructure/utils/query-params-utils.js';
+import * as queryParamsUtils from '../../../../../src/shared/infrastructure/utils/query-params-utils.js';
 
 describe('Unit | Application | learner-list-controller', function () {
   describe('#findPaginatedFilteredParticipants', function () {
