+++ api/tests/unit/application/session/session-controller_test.js
@@ -7,7 +7,7 @@ import { UserLinkedToCertificationCandidate } from '../../../../lib/domain/event
 import { SessionPublicationBatchResult } from '../../../../lib/domain/models/index.js';
 import { logger } from '../../../../src/shared/infrastructure/utils/logger.js';
 import { SessionPublicationBatchError } from '../../../../lib/application/http-errors.js';
-import * as queryParamsUtils from '../../../../lib/infrastructure/utils/query-params-utils.js';
+import * as queryParamsUtils from '../../../../src/shared/infrastructure/utils/query-params-utils.js';
 import { getI18n } from '../../../tooling/i18n/i18n.js';
 
 describe('Unit | Controller | sessionController', function () {
